refactor -  getEndBeat and getStartBeat have them own files
